+++ x/dclauth/keeper/account.go
@@ -128,7 +128,7 @@ func (k Keeper) HasRole(ctx sdk.Context, addr sdk.AccAddress, roleToCheck types.
 }
 
 // Check if account has vendorID association.
-func (k Keeper) HasVendorID(ctx sdk.Context, addr sdk.AccAddress, vid uint64) bool {
+func (k Keeper) HasVendorID(ctx sdk.Context, addr sdk.AccAddress, vid int32) bool {
 	account, found := k.GetAccountO(ctx, addr)
 
 	if !found {
